Move cmd/ingest to internal/ingest
