+++ internal/ingest/fakecs_test.go
@@ -1,4 +1,4 @@
-package main
+package ingest
 
 import (
 	"crypto/sha512"
