+++ internal/ingest/testcs_test.go
@@ -1,4 +1,4 @@
-package main
+package ingest
 
 import (
 	"bytes"
